[detailed-metrics] extract TSolomonCounters into NPrivate::TAggregatedTabletCounters
